Revised methods for title slides
Found we only had methods for some sizes, and didn't have a consistent
way to fiddle the formatting when you wanted *almost* that slide with a
little difference.

Also has a hacky vertical alignment that works ok for single lines of
text. Worth improving later, but gets it done for now.